Set time for current day to 12 to avoid tz issue
====  Tech Notes  ====
The prevously used datetime for getting current day: YYYY-MM-DD 00:00
caused problems for timezones on the + side of GMT due to the
recalulation to GMT. Sweden for example is GMT +1 wich when recalculated
yields the datetime YYYY-MM-DD-1 23:00 (i.e. 23:00 the previous day).
OWM forcast doesn't work for previous days. and causes an error.

Setting the time to 12 should work in all cases and since for both
positive and  negative timezones.