Fix OWM language selection
self.lang refers to the language set in mycroft.conf, a BCP-47 language tag. The first part of the tag refers to the language
which should have preference if supported by OWM. The second part of the tag is the region, which should not be used as a language
but might still be a better fallback than english.