Move dialog data translation to build location
This updates the way the translation of data elements for dialog
redering is performed. It moves the translation to where the data is
generated. By keeping translation close to the building it is easy to
see which keys exists in the data reducing the risk of making mistakes.

It also makes the dialog more self-contained.