Fix sunrise/set sometimes reporting incorrect time
The datetime localizer was going through a naive state which
was causing issues in very specific cases. This now creates
a tz aware object directly from the epoch based timestamp.

Pytz has also been added as a requirement with no version
limitation as this is likely defined by another component.